guild model fix
