+++ src/base/GCommandsDispatcher.js
@@ -22,7 +22,7 @@ class GCommandsDispatcher {
     async setGuildPrefix(prefix, guildId) {
         if(!this.client.database || !this.client.database.working) return this.client.prefix;
         if(this.client.database.type = "mongodb") {
-            var guildSettings = require('./utils/models/guild')
+            var guildSettings = require('../util/models/guild')
 
             const settings = await guildSettings.findOne({ id: guildId})
             if(!settings) {
@@ -56,7 +56,7 @@ class GCommandsDispatcher {
     async getGuildPrefix(guildId) {
         if(!this.client.database || !this.client.database.working) return this.client.prefix;
         if(this.client.database.type = "mongodb") {
-            var guildSettings = require('./utils/models/guild')
+            var guildSettings = require('../util/models/guild')
 
             const settings = await guildSettings.findOne({ id: guildId})
             if(!settings) {
