--- src/managers/CommandLoad.js
@@ -1,1 +1,0 @@
-/* now in GCommands.js */
